doc: Update Exported Identifiers RDR (#30)
The _Exported Names_ RDR has been renamed to _Exported Identifiers_ and the
status has been set back to `Drafting` status. We'll move RDR's into `Accepted`
status when we make a release. The examples in the document have been improved,
as well as calling out an important issue with the chosen design.