build: Run Dialyzer in CI (#43)
Dialyzer emits a stream of warnings related to code outside the project when it
builds a PLT, the first time it runs. A new `prime-dialyzer` Make target pipes
dialyzer output to `/dev/null` and ignores the exit code so that a subsequent
run of Dialyzer only reports on issues in the project. The `ci` Make target now
invokes `prime-dialyzer` and `dialyzer` Make targets to ensure that Dialyzer
errors cause CI builds to fail.