compiler: Annotate and typecheck forms for function apply expressions (#45)
A new `rufus_form:annotate/3` function provides a clean way to set or update
context values in a Rufus form. A new `rufus_scope:globals/1` function takes a
list of Rufus forms and transforms them into a map that's used when resolving
global names. A new `rufus_type/resolve/2` function returns the type information
associated with a Rufus form, or infers it if it isn't available. A new
`rufus_typecheck_apply` module defines a compiler stage that performs
typechecking to ensure that function application is correct: that the function
exists, with the right arity, with matching argument types, and with no
ambiguity as to which function declarations match the apply callsite. It returns
annotate forms with every `apply` form receiving an inferred `type`.