compiler: All forms are annotated with type information (#70)
`rufus_expr:typecheck_and_annotate/1` ensures that every form has a `type`
annotation. A sanity check is run after typechecking and annotations and an
`{error, sanity_check, Data}` error triple is returned if any form is missing a
`type` annotation. This change was made to remove brittle code in `rufus_erlang`
that had to inspect a `locals` annotation to fetch the type. Ensuring every form
has type information allows us to constrain the use of `locals` to the
`rufus_expr` code path.