compiler: Parse typed list literals (#86)
`rufus_tokenize:string/1` inserts semicolons after `}` characters when they're
the last character on a line or the last character in the source text, instead
of inserting the semicolon before the `}`. The parser is updated to parse list
literals that are now spelled like `list[int]{1, 2, 3}` instead of `[1, 2, 3]`.
In part, this is to resolve an issue with an unknown type when an empty list
literal was specified. Now, empty lists include their type, such as
`list[string]{}`. A new `rufus_form:make_literal/4` can be used to create a list
literal with a correctly formed list type form.